requirements.txt
beautifulsoup4==4.12.2
gym==0.26.2
joblib==1.2.0
langchain==0.0.205
numpy==1.24.3
openai==0.27.8
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
PyYAML==6.0
PyYAML==6.0
regex==2022.7.9
selenium==4.10.0
setuptools==67.8.0
tiktoken==0.4.0
tqdm==4.65.0
urllib3==1.26.15
